--- src/style.css
@@ -1,115 +1,0 @@
-*{
-    font-family:system-ui, -apple-system, BlinkMacSystemFont, 'Segoe UI', Roboto, Oxygen, Ubuntu, Cantarell, 'Open Sans', 'Helvetica Neue', sans-serif;
-}
-
-
-body{
-    display: flex;
-    flex-direction: column;
-    justify-content: center;
-    align-items: center;
-    height: 100vh;
-    gap: 30px;
-}
-.dynamic_visual_scoring_calculator {
-    display: flex;
-    flex-direction: row;
-    flex-wrap: wrap;
-    justify-content: center;
-    justify-items: center;
-    align-items: center;
-    align-content: center;
-    gap: 100px;
-}
-.dvsc_output_svg{
-    height: 200px;
-    width: 200px;
-    position: relative;
-}
-.dvsc_output_svg *{
-    transition: all 0.4s ease-out;
-}
-#dvsc_score_indicator{
-    cx: 50;
-    cy: 50;
-    r: 40;
-    stroke-width: 4px;
-    stroke-linecap: round;
-    fill: none;
-}
-#dvsc_score_text{
-    text-anchor: middle;
-    font-family:'Courier New', Courier, monospace;
-    font-size: 32px;
-    font-weight: bold;
-    dominant-baseline: middle;
-}
-
-.fraction {
-    cx: 50;
-    cy: 50;
-    r: 46;
-    stroke-width: 3px;
-    stroke-linecap: round;
-    fill: none;
-    transform-origin: center;
-}
-
-.fraction_bg{
-    opacity: 0.5;
-}
-
-.fraction_bar{
-    opacity: 1;
-}
-
-
-.tab, .tab > p{
-    tab-size: 4;
-    margin: 0;
-    text-align: end;
-    white-space: nowrap;
-    word-break: keep-all;
-}
-
-
-tr {
-    display: grid;
-    grid-gap: 45px;
-    grid-auto-flow: column;
-    grid-template-columns: 90px 65px 216px 164px 216px 110px;
-    justify-content: space-between;
-}
-
-td {
-    display: flex;
-    justify-content: space-between;
-    align-items: center;
-    gap: 15px;
-    flex-wrap: nowrap;
-    text-align: end;
-}
-
-th>p {
-    text-align: left;
-}
-thead>tr>th:nth-child(1){
-    text-align: left;
-}
-table>tbody>tr>td:nth-child(2), table>tbody>tr>td:nth-child(6){
-    justify-content: center;
-}
-
-
-.colors{
-    	color: #ff0000;
-    	color: #ff7b00;
-        color: #ffbb00;
-    	color: #00dfba;
-    	color: #00d5e4;
-    	color: #127efa;
-    	color: #8921ff;
-    	color: #d500e9;
-
-        background: linear-gradient(hsl(115, 94%, 45%),#ffff);
-}
